+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v0.2.1
+1, fix audio augments  
+2, update audio document  
+3, change audio type from numpy to torch  
+4, change sampling_rate to sample_rate
+
 ### v0.1.3
 1, fix requirements and MANIFEST  
 2, add extra info for datasource and document exported dict